+++ decidim-proposals/app/permissions/decidim/proposals/admin/permissions.rb
@@ -32,6 +32,12 @@ module Decidim
           # Every user allowed by the space can import proposals from another_component
           allow! if permission_action.subject == :proposals && permission_action.action == :import
 
+          # Every user allowed by the space can merge proposals to another component
+          allow! if permission_action.subject == :proposals && permission_action.action == :merge
+
+          # Every user allowed by the space can split proposals to another component
+          allow! if permission_action.subject == :proposals && permission_action.action == :split
+
           if permission_action.subject == :participatory_texts && participatory_texts_are_enabled?
             # Every user allowed by the space can import participatory texts to proposals
             allow! if permission_action.action == :import