+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Decidim::User.where(**search for old subscribed users**).update(newsletter_notif
 - **decidim-core**: Added metrics visualization for Assemblies, ParticipatoryProcesses, Results (Accountability), Comments, and Meetings [\#36042283](https://github.com/decidim/decidim/pull/4228)
 - **decidim-core**: Let admins and creators edit the user group profile [\#4283](https://github.com/decidim/decidim/pull/4283)
 - **decidim-core**: User groups can also have badges. [\#4310](https://github.com/decidim/decidim/pull/4310)
+- **decidim-proposals**: Merge and split proposals [\#4360](https://github.com/decidim/decidim/pull/4360)
 
 **Changed**: